Move packages_autoroller and delete conductor directory
