--- dev/conductor/README.md
@@ -1,8 +1,0 @@
-# Flutter Conductor
-
-This directory contains the skeletal remains of the `conductor` tool, which was
-used to semi-automate releases. It still has a functional tool
-(`packages_autoroller`) which is in the process of being moved to a dedicated
-directory.
-
-Questions? Ask @matanlurey.
