--- dev/conductor/bin/packages_autoroller
@@ -1,47 +1,0 @@
-#!/usr/bin/env bash
-# Copyright 2014 The Flutter Authors. All rights reserved.
-# Use of this source code is governed by a BSD-style license that can be
-# found in the LICENSE file.
-
-set -euo pipefail
-
-# Needed because if it is set, cd may print the path it changed to.
-unset CDPATH
-
-# On Mac OS, readlink -f doesn't work, so follow_links traverses the path one
-# link at a time, and then cds into the link destination and find out where it
-# ends up.
-#
-# The returned filesystem path must be a format usable by Dart's URI parser,
-# since the Dart command line tool treats its argument as a file URI, not a
-# filename. For instance, multiple consecutive slashes should be reduced to a
-# single slash, since double-slashes indicate a URI "authority", and these are
-# supposed to be filenames. There is an edge case where this will return
-# multiple slashes: when the input resolves to the root directory. However, if
-# that were the case, we wouldn't be running this shell, so we don't do anything
-# about it.
-#
-# The function is enclosed in a subshell to avoid changing the working directory
-# of the caller.
-function follow_links() (
-  cd -P "$(dirname -- "$1")"
-  file="$PWD/$(basename -- "$1")"
-  while [[ -h "$file" ]]; do
-    cd -P "$(dirname -- "$file")"
-    file="$(readlink -- "$file")"
-    cd -P "$(dirname -- "$file")"
-    file="$PWD/$(basename -- "$file")"
-  done
-  echo "$file"
-)
-
-PROG_NAME="$(follow_links "${BASH_SOURCE[0]}")"
-BIN_DIR="$(cd "${PROG_NAME%/*}" ; pwd -P)"
-REPO_DIR="$BIN_DIR/../../.."
-DART_BIN="$REPO_DIR/bin/dart"
-FLUTTER_BIN="$REPO_DIR/bin/flutter"
-
-# Ensure pub get has been run in the repo before running the conductor
-(cd "$REPO_DIR/dev/conductor/core"; "$FLUTTER_BIN" pub get 1>&2)
-
-exec "$DART_BIN" --enable-asserts "$REPO_DIR/dev/conductor/core/bin/packages_autoroller.dart" "$@"
