--- dev/conductor/core/analysis_options.yaml
@@ -1,12 +1,0 @@
-include: ../../analysis_options.yaml
-
-analyzer:
-  exclude:
-    # Ignore protoc generated files
-    - "lib/src/proto/*"
-
-linter:
-  rules:
-    avoid_catches_without_on_clauses: true
-    prefer_relative_imports: true
-    unawaited_futures: true
