+++ dev/packages_autoroller/lib/src/git.dart
