+++ dev/packages_autoroller/lib/src/stdio.dart
