+++ dev/packages_autoroller/pubspec.yaml
@@ -1,4 +1,4 @@
-name: conductor_core
+name: packages_autoroller
 description: Flutter Automated Release Tool
 
 publish_to: none
@@ -11,7 +11,6 @@ resolution: workspace
 dependencies:
   args: any
   file: any
-  fixnum: any
   meta: any
   path: any
   platform: any
